Address PR review feedback on config-driven form skeleton
Rename FormConfig.entities to FormConfig.fields — "fields" is the
natural form-domain term, "entities" was an internal-only concept that
leaked into the public API.

Split config types into config-types.ts to separate the config-driven
engine types (FieldConfig, LayoutConfig, FormConfig, renderers) from
the form runtime types (FormProps, FormState, FormApi). This keeps
form-types.ts focused on the component API consumers already know.

Restore @example blocks on FormProps.render and FormProps.footer that
were removed in the initial skeleton commit — existing JSDoc should
not be stripped as part of unrelated changes.

Add JSDoc to FormProvider, FormContextType, and useFormProvider matching
the documentation pattern established by CellProvider.

Remove sandbox changes from this branch — proof-of-life demos belong
on the final stack branch, not the skeleton.
