+++ javascript/packages/core/components/form/config-driven-form.tsx
@@ -1,7 +1,8 @@
 import { Form } from '#core/components/form/form';
 import { LayoutItemList } from '#core/components/form/layout/layout-item-list';
 
-import type { FormConfig, FormData } from '#core/components/form/types/form-types';
+import type { FormConfig } from '#core/components/form/types/config-types';
+import type { FormData } from '#core/components/form/types/form-types';
 
 type ConfigDrivenFormProps = {
   config: FormConfig;
@@ -12,7 +13,7 @@ type ConfigDrivenFormProps = {
 export function ConfigDrivenForm({ config, onSubmit, initialValues }: ConfigDrivenFormProps) {
   return (
     <Form onSubmit={onSubmit} initialValues={initialValues}>
-      <LayoutItemList items={config.layout} entities={config.entities} />
+      <LayoutItemList items={config.layout} fields={config.fields} />
     </Form>
   );
 }
