+++ javascript/packages/core/components/form/hooks/use-layout-renderer.ts
@@ -1,7 +1,7 @@
 import { LAYOUT_RENDERERS } from '#core/components/form/constants';
 import { useFormProvider } from '#core/providers/form-provider/use-form-provider';
 
-import type { LayoutRenderer } from '#core/components/form/types/form-types';
+import type { LayoutRenderer } from '#core/components/form/types/config-types';
 
 export function useLayoutRenderer(type: string): LayoutRenderer | undefined {
   const formContext = useFormProvider();
