+++ javascript/packages/core/components/form/layout/layout-item-list.tsx
@@ -1,22 +1,22 @@
 import { SchemaField } from '#core/components/form/fields/schema-field';
 import { LayoutItemRenderer } from './layout-item-renderer';
 
-import type { FieldConfig, LayoutItem } from '#core/components/form/types/form-types';
+import type { FieldConfig, LayoutItem } from '#core/components/form/types/config-types';
 
 export function LayoutItemList({
   items,
-  entities,
+  fields,
 }: {
   items: LayoutItem[];
-  entities: Record<string, FieldConfig>;
+  fields: Record<string, FieldConfig>;
 }) {
   return (
     <>
       {items.map((item, index) =>
         typeof item === 'string' ? (
-          <SchemaField key={item} fieldPath={item} config={entities[item]} />
+          <SchemaField key={item} fieldPath={item} config={fields[item]} />
         ) : (
-          <LayoutItemRenderer key={index} config={item} entities={entities} />
+          <LayoutItemRenderer key={index} config={item} fields={fields} />
         )
       )}
     </>
