+++ javascript/packages/core/components/views/sandbox/sandbox.tsx
@@ -1,42 +1,19 @@
 import { Block } from 'baseui/block';
 import { HeadingXXLarge } from 'baseui/typography';
 
-import { ConfigDrivenForm } from '#core/components/form/config-driven-form';
+import { StringField } from '#core/components/form/fields/string/string-field';
+import { Form } from '#core/components/form/form';
 import { MainViewContainer } from '#core/components/views/main-view-container';
-
-import type { FormConfig } from '#core/components/form/types/form-types';
-
-const SAMPLE_FORM_CONFIG: FormConfig = {
-  entities: {
-    'spec.title': { type: 'string', label: 'Title', required: true, placeholder: 'Enter a title' },
-    'spec.description': {
-      type: 'string',
-      label: 'Description',
-      placeholder: 'Enter a description',
-    },
-    'spec.tags': { type: 'string', label: 'Tags', multi: true, placeholder: 'Add a tag' },
-  },
-  layout: [
-    {
-      type: 'group',
-      label: 'General',
-      items: ['spec.title', 'spec.description'],
-    },
-    {
-      type: 'group',
-      label: 'Metadata',
-      items: ['spec.tags'],
-    },
-  ],
-};
 
 export function Sandbox() {
   return (
     <MainViewContainer>
       <HeadingXXLarge>Component Sandbox</HeadingXXLarge>
-      <Block marginBottom="24px">Config-driven form proof-of-life.</Block>
-      <Block width="600px">
-        <ConfigDrivenForm config={SAMPLE_FORM_CONFIG} onSubmit={() => undefined} />
+      <Block marginBottom="24px">This is a sandbox for testing WIP components and features.</Block>
+      <Block width="400px">
+        <Form onSubmit={() => undefined}>
+          <StringField name="tags" label="Tags" multi placeholder="Add a tag" />
+        </Form>
       </Block>
     </MainViewContainer>
   );
