+++ javascript/packages/core/providers/form-provider/form-provider.tsx
@@ -4,6 +4,27 @@ import { FormContext } from './form-context';
 
 import type { FormContextType } from './types';
 
+/**
+ * @description
+ * Provider component that allows consumers to register custom field renderers
+ * and layout renderers. Custom renderers are checked before falling back to
+ * built-in renderers.
+ *
+ * @example
+ * ```tsx
+ * const fieldRenderers = {
+ *   'hive-select': HiveSelectField,
+ * };
+ *
+ * const layoutRenderers = {
+ *   tabs: TabsLayoutRenderer,
+ * };
+ *
+ * <FormProvider renderers={fieldRenderers} layouts={layoutRenderers}>
+ *   <ConfigDrivenForm config={formConfig} onSubmit={handleSubmit} />
+ * </FormProvider>
+ * ```
+ */
 export const FormProvider = ({
   children,
   renderers = {},
